move websocket-clock folder to another place and change require links
