+++ src/views/admin.pug
@@ -29,7 +29,7 @@ block content
                 button.btn.btn-dark(type='submit') Save dayOfWeek
 
               form#formTable(method='POST' action='/admin/submit')
-                button.btn.btn-dark(type='reset' style='width: 100%' name="adminsubmit", id="adminsubmit") Save
+                button.btn.btn-dark(type='submit' style='width: 100%' name="adminsubmit", id="adminsubmit") Save
                 .form-row
                     .col-1.text-right
                       | #
